Wrangle the OpenBSD branch into shape as starting point:
• Revert “Use the new libc uid_from_user() and gid_from_group()
  instead of the pax-specific functions in cache.c” (revisit later)
• fix MAX_TIME_T (plain wrong), mirtoconf later
• drop all NOCPIO
• Revert “Replace name_{uid,gid}() with the libc routines
  user_from_uid() and group_from_gid()”

+++ pax.h
@@ -259,4 +259,4 @@ typedef struct oplist {
 #define _PAX_		1
 #define _TFILE_BASE	"paxXXXXXXXXXX"
 #define MAX_TIME_T	(sizeof(time_t) == sizeof(long long) ? \
-			    LLONG_MAX : INT_MAX)
+			    LLONG_MAX : LONG_MAX)
